Updating settings, adding converter settings
