+++ src/renderer/components/Settings/styles.ts
@@ -1,0 +1,35 @@
+export const styles = {
+    main: {
+        height: '100%',
+        overflow: 'overlay',
+    },
+    tabPanel: {
+        padding: 4,
+        flex: '1 1 99%',
+    },
+    buttonContainer: {
+        padding: 4,
+        display: 'flex',
+        justifyContent: 'flex-end',
+    },
+    paper: {
+        height: '100%',
+        backgroundColor: 'grey.50',
+    },
+    helperText: {
+        paddingLeft: 2,
+        margin: 0,
+        mt: 0,
+    },
+    tab: {
+        background:
+            'radial-gradient(circle farthest-corner at bottom center,#eeeeee,#e5e4e4)',
+    },
+    inputField: {
+        maxWidth: '600px',
+    },
+    alert: {
+        mt: 1,
+        mb: 2,
+    },
+};
