+++ src/renderer/redux/initialState.ts
@@ -9,6 +9,7 @@ export const settings: ISettings = {
         maxColWidth: 100,
         dateFormat: 'ISO8601',
         roundNumbers: false,
+        maxPrecision: 8,
         copyFormat: 'tab',
     },
     converter: {
@@ -19,52 +20,13 @@ export const settings: ISettings = {
             'MMDDYY',
             'YYMMDD',
             'B8601DA',
-            'B8601DN',
             'E8601DA',
-            'E8601DN',
-            'NENGO',
-            'WORDDATE',
-            'WORDDATX',
-            'PDJULG',
-            'PDJULI',
-            'NLDATE',
-            'NLDATEL',
         ],
-        timeFormats: [
-            'TIME',
-            'HHMM',
-            'HOUR',
-            'MMSS',
-            'E8601TM',
-            'B8601TM',
-            'TIMEAMPM',
-            'B8601TX',
-            'B8601TZ',
-            'E8601TX',
-            'E8601TZ',
-            'B8601LZ',
-            'E8601LZ',
-            'NLTIME',
-            'NLTIMAP',
-        ],
-        datetimeFormats: [
-            'DATETIME',
-            'E8601DT',
-            'B8601DT',
-            'MDYAMPM',
-            'DATEAMPM',
-            'B8601DX',
-            'B8601DZ',
-            'B8601LX',
-            'E8601DX',
-            'E8601DZ',
-            'E8601LX',
-            'NLDATM',
-            'NLDATML',
-        ],
+        timeFormats: ['TIME', 'HHMM', 'MMSS', 'E8601TM', 'B8601TM'],
+        datetimeFormats: ['DATETIME', 'E8601DT', 'B8601DT'],
         convertSuffixDt: false,
         convertSuffixTm: false,
-        convertSuffixDtTm: false,
+        convertSuffixDtm: false,
     },
     other: {
         checkForUpdates: true,
